add `canonical` option

+++ ansible_collections/bofzilla/purelymail/plugins/modules/routing_rules.py
@@ -21,12 +21,19 @@ description:
     - (Any address starting with) → C(match_user="<anything you want>", prefix=True, catchall=False)
     - (The exact address) → C(match_user="<anything you want>", prefix=False, catchall=False)
   - When using presets, you can provide the preset name in the rule instead of manually setting prefix/catchall/match_user.
+  - Optionally, `canonical` can be set to true to remove any rules not explicitly defined in the input list.
 
 options:
   api_token:
     description: Purelymail API token
     required: true
     type: str
+
+  canonical:
+    description: If true, remove any existing rules not specified in `rules`
+    type: bool
+    required: false
+    default: false
 
   rules:
     description: List of routing rules to apply
@@ -96,6 +103,7 @@ RETURN = r""""""
 module_spec = dict(
 	argument_spec=dict(
 		api_token=dict(type="str", required=True, no_log=True),
+		canonical=dict(type="bool", required=False, default=False),
 		rules=dict(
 			type="list",
 			required=True,
@@ -146,14 +154,17 @@ def main():
 		extra_rules = [er.id for er in existing_rules if not any(r.matches(er) for r in rules)]
 		missing_rules = [r for r in rules if not any(r.matches(er) for er in existing_rules)]
 
-		if len(extra_rules) == 0 and len(missing_rules) == 0:
+		would_change = (module.params["canonical"] and extra_rules) or missing_rules
+
+		if not would_change:
 			module.exit_json(changed=False)
 
 		if module.check_mode:
 			module.exit_json(changed=True)
 
-		for id in extra_rules:
-			client.delete_route(DeleteRoutingRequest(id))
+		if module.params["canonical"]:
+			for id in extra_rules:
+				client.delete_route(DeleteRoutingRequest(id))
 
 		for rule in missing_rules:
 			client.create_route(rule)
